requirements.txt
# Core ML and Data Processing
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Streaming and Real-time Processing
kafka-python==2.0.2
apache-kafka==3.4.0
redis==4.6.0

# Data Visualization and Monitoring
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# API and Web Framework
fastapi==0.100.1
uvicorn==0.23.2
pydantic==2.0.3

# Database and Storage
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pymongo==4.4.0

# Utilities and Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.4.2

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Docker and Deployment
docker==6.1.3
docker-compose==1.29.2

# Monitoring and Logging
prometheus-client==0.17.1
structlog==23.1.0

# Additional ML Libraries
isolation-forest==0.2.2
pyod==1.0.8 